vehicle_monitoring_system/requirements.txt
aio-pika==7.2.0
aiormq==6.2.3
asyncio==3.4.3
autopep8==1.6.0
idna==3.3
motor==3.0.0
multidict==6.0.2
pamqp==3.1.0
pycodestyle==2.8.0
pymongo==4.1.1
python-dotenv==0.20.0
toml==0.10.2
yarl==1.7.2
